fix(structure): organize code into directories
