--- src/DataDictionary/graph/DataDictionaryGraph.css
@@ -1,58 +1,0 @@
-:root {
-  --opacity--not-highlighted: 0.1;
-}
-
-.data-dictionary-graph__node {
-}
-
-.data-dictionary-graph__node-rect {
-  fill: var(--g3-color__white);
-  stroke-width: 2px;
-}
-
-.data-dictionary-graph__node:hover .data-dictionary-graph__node-rect {
-  fill: var(--g3-color__highlight-orange);
-}
-
-.data-dictionary-graph__link {
-  fill: none;
-  stroke: black;
-}
-
-.data-dictionary-graph__link--required {
-  stroke: var(--g3-color__highlight-orange);;
-}
-
-.data-dictionary-graph__link--highlighted {
-  stroke-width: 3px;
-}
-
-.data-dictionary-graph__link--not-highlighted {
-  opacity: var(--opacity--not-highlighted)
-}
-
-.data-dictionary-graph__node--not-highlighted {
-  opacity: var(--opacity--not-highlighted);
-  cursor: none;
-  pointer-events: none;
-}
-
-.data-dictionary-graph__node--clickable {
-  cursor: pointer;
-}
-
-.data-dictionary-graph__node--not-clickable {
-  cursor: default;
-}
-
-.data-dictionary-graph__text {
-  font-weight: 600;
-  fill: var(--g3-color__black);
-}
-
-.data-dictionary-graph__node:hover .data-dictionary-graph__text {
-  fill: var(--g3-color__white);
-}
-
-
-
