+++ src/DataDictionary/table/DataDictionaryPropertyTable/DataDictionaryPropertyTable.css
